fix: json 함수를 json.ts로 위치 이동
