+++ src/shared/cache/json.ts
@@ -1,5 +1,44 @@
-import { getCacheData, saveFile } from "../../features/blog/services/notion";
 import { safeClone } from "../../utils/common";
+import path from "path";
+import fs from "fs";
+
+export const saveFile = (src: string, slug: string, json: any) => {
+  try {
+    const outDir = path.join(process.cwd(), src);
+    if (!fs.existsSync(outDir)) fs.mkdirSync(outDir, { recursive: true });
+    const extension = slug.split(".").at(-1);
+    if (!extension) return new Error("파일 확장자를 작성하지 않았습니다.");
+    if (extension === "js") {
+      fs.writeFileSync(path.join(outDir, slug), json);
+    } else {
+      fs.writeFileSync(path.join(outDir, slug), JSON.stringify(json, null, 2));
+    }
+    return { message: "success", id: src + slug };
+  } catch (error) {
+    return { message: "fail: " + error };
+  }
+};
+
+export const getCacheData = (src: string) => {
+  const filePath = path.join(process.cwd(), src);
+  if (fs.existsSync(filePath)) {
+    const fileContent = fs.readFileSync(filePath, "utf-8");
+    if (src.endsWith(".js")) {
+      return fileContent;
+    }
+    return JSON.parse(fileContent);
+  }
+  return null;
+};
+
+export const deletedCacheData = (src: string) => {
+  const filePath = path.join(process.cwd(), src);
+  if (fs.existsSync(filePath)) {
+    const fileContent = fs.rmdirSync(filePath);
+    return { message: "success" };
+  }
+  return null;
+};
 
 export const updateJSONFile = <T>(
   fullPath: string,
